server.js szétbontás
